Cleanup of ConfigValidator tests
* pulled out ConfigValidatorEvictionConfigTest
* pulled out ConfigValidatorNearCacheConfigTest
* fixed some tests names in ConfigValidatorTest
